Merged count function and the reply function.
Makes the code a little bit cleaner, and also I found a variable msgReplied (renamed from userMatch) that I'm not sure does anything. I think there should be a line for it to be equal to false right when a message is received otherwise there may be some double replied messages, but I haven't seen anything like that. So I'm really at a loss at what this variable does (might even be useless).